fixup! refactor(router): Add provider for integrating with Navigation API and Location shim
