+++ packages/common/src/location/location.ts
@@ -6,12 +6,11 @@
  * found in the LICENSE file at https://angular.dev/license
  */
 
-import {inject, Injectable, OnDestroy, ɵɵinject, DestroyRef} from '@angular/core';
+import {Injectable, OnDestroy, ɵɵinject} from '@angular/core';
 import {Subject, SubscriptionLike} from 'rxjs';
 
 import {LocationStrategy} from './location_strategy';
 import {joinWithSlash, normalizeQueryParams, stripTrailingSlash} from './util';
-import {PlatformNavigation} from '../navigation/platform_navigation';
 
 /** @publicApi */
 export interface PopStateEvent {
@@ -302,61 +301,6 @@ export class Location implements OnDestroy {
    * @returns The URL string, modified if needed.
    */
   public static stripTrailingSlash: (url: string) => string = stripTrailingSlash;
-}
-
-@Injectable()
-export class NavigationAdapterForLocation extends Location {
-  private readonly navigation = inject(PlatformNavigation);
-  private readonly destroyRef = inject(DestroyRef);
-
-  constructor() {
-    super(inject(LocationStrategy));
-
-    const currentEntryChangeListener = () => {
-      this._notifyUrlChangeListeners(this.path(true), this.getState());
-    };
-    this.navigation.addEventListener('currententrychange', currentEntryChangeListener);
-    this.destroyRef.onDestroy(() => {
-      this.navigation.removeEventListener('currententrychange', currentEntryChangeListener);
-    });
-  }
-
-  override getState(): unknown {
-    return this.navigation.currentEntry?.getState();
-  }
-
-  override replaceState(path: string, query: string = '', state: any = null): void {
-    const url = this.prepareExternalUrl(path + normalizeQueryParams(query));
-    // Use navigation API consistently for navigations. The "navigation API state"
-    // field has no interaction with the existing "serialized state" field, which is what backs history.state
-    this.navigation.navigate(url, {state, history: 'replace'});
-  }
-
-  override go(path: string, query: string = '', state: any = null): void {
-    const url = this.prepareExternalUrl(path + normalizeQueryParams(query));
-    // Use navigation API consistently for navigations. The "navigation API state"
-    // field has no interaction with the existing "serialized state" field, which is what backs history.state
-    this.navigation.navigate(url, {state, history: 'push'});
-  }
-
-  // Navigation.back/forward differs from history in how it traverses the joint session history
-  // https://github.com/WICG/navigation-api?tab=readme-ov-file#correspondence-with-the-joint-session-history
-  override back() {
-    this.navigation.back();
-  }
-
-  override forward() {
-    this.navigation.forward();
-  }
-
-  override onUrlChange(fn: (url: string, state: unknown) => void): VoidFunction {
-    this._urlChangeListeners.push(fn);
-
-    return () => {
-      const fnIndex = this._urlChangeListeners.indexOf(fn);
-      this._urlChangeListeners.splice(fnIndex, 1);
-    };
-  }
 }
 
 export function createLocation() {
